fix some warnings. Move one orphan instance from Reference to AssumpChunk.
