+++ code/drasil-lang/Language/Drasil/Reference.hs
@@ -4,7 +4,6 @@ module Language.Drasil.Reference(makeRef2, makeRef2S, makeCite,
 
 import Control.Lens ((^.))
 
-import Language.Drasil.Chunk.AssumpChunk as A (AssumpChunk)
 import Language.Drasil.Chunk.Citation as Ci (citeID, Citation)
 import Language.Drasil.Chunk.Concept (ConceptInstance, sDom)
 import Language.Drasil.Chunk.DataDefinition (DataDefinition)
@@ -13,22 +12,13 @@ import Language.Drasil.Chunk.InstanceModel (InstanceModel)
 import Language.Drasil.Chunk.Theory (TheoryModel)
 import Language.Drasil.Classes.Core (HasUID(uid), HasRefAddress(getRefAdd),
   HasShortName(shortname))
-import Language.Drasil.Classes (ConceptDomain(cdom), abrv)
+import Language.Drasil.Classes (ConceptDomain(cdom), abrv, Referable(refAdd, renderRef))
 import Language.Drasil.Document (Section(Section))
 import Language.Drasil.Document.Core (LabelledContent(..), RawContent(..))
 import Language.Drasil.Label.Type (LblType(RP,Citation), IRefProg,
   prepend, name, raw, (+::+), defer)
 import Language.Drasil.RefProg (Reference(Reference))
 import Language.Drasil.Sentence (Sentence(Ref))
-
-class HasUID s => Referable s where
-  refAdd    :: s -> String  -- The referencing address (what we're linking to).
-                            -- Only visible in the source (tex/html).
-  renderRef :: s -> LblType -- alternate
-
-instance Referable AssumpChunk where
-  refAdd    x = getRefAdd x
-  renderRef l = RP (prepend $ abrv l) (refAdd l)
 
 instance Referable Section where
   refAdd    (Section _ _ lb ) = getRefAdd lb
